+++ src/widgets/docker/component.jsx
@@ -1,10 +1,9 @@
 import { useTranslation } from "next-i18next/pages";
 import useSWR from "swr";
 
-import { calculateCPUPercent, calculateThroughput, calculateUsedMemory } from "./stats-helpers";
-
 import Block from "components/services/widget/block";
 import Container from "components/services/widget/container";
+import { calculateCPUPercent, calculateThroughput, calculateUsedMemory } from "utils/docker/stats-helpers";
 
 export default function Component({ service }) {
   const { t } = useTranslation();
